quantity function cleanup done
